Moved routes for habits, chores, expenses, and balance to separate files
