+++ index.js
@@ -3,10 +3,13 @@ const ecstatic =  require('ecstatic');
 const loader = require('./modules/dataLoader.js');
 const Router = require('./modules/router.js')
 const templateHandler = require('./modules/templateHandler.js');
-const Habit = require('./modules/sharedLibs/habit.js');
 const Balance = require('./modules/sharedLibs/balance.js');
-const Expense = require('./modules/sharedLibs/expense.js');
 require('./modules/sharedLibs/sharedLib.js');
+
+const helpers = require('./modules/routeHelpers');
+const simpleGET = helpers.simpleGET;
+const respondBadReq = helpers.respondBadReq;
+const validateRequest = helpers.validateRequest;
 
 let router = new Router();
 
@@ -58,249 +61,12 @@ router.add('GET', /^\/shared-lib\/(.+)$/, function (request, response, filename)
   sharedLibServer(request, response);
 });
 
-/* Return a list of all habits in JSON for any client to consume */
-router.add('GET', /^\/all-habits$/, function (request, response) {
-  simpleGET(loader.allHabits(), response);
-});
-
-/* Return a list of active habits in JSON for any client to consume */
-router.add('GET', /^\/active-habits$/, function (request, response) {
-  simpleGET(loader.activeHabits(), response);
-});
-
-/* Return a list of active habits in JSON for any client to consume */
-router.add('GET', /^\/archived-habits$/, function (request, response) {
-  simpleGET(loader.archivedHabits(), response);
-});
-
-/* Return a list of all expenses in JSON for any client to consume */
-router.add('GET', /^\/all-expenses$/, function (request, response) {
-  simpleGET(loader.allExpenses(), response);
-});
-
-router.add('GET', /^\/habits-left\/(\d{4}-\d{2}-\d{2})$/,
-  (request, response, dateStr) => {
-    /* TODO: Verify the date string */
-    simpleGET(loader.habitsLeft(dateStr), response);
-  }
+router.routes = router.routes.concat(
+  require('./routes/habitRoutes').routes,
+  require('./routes/expenseRoutes').routes,
+  require('./routes/choreRoutes').routes,
+  require('./routes/balanceRoutes').routes
 );
-
-/* Get the info for a single habit given the habit's document id */
-router.add('GET', /^\/habit\/(\w+)$/, function (request, response, id) {
-  if (!id) respondBadReq(response, 'request must contain habit ID in path');
-  else simpleGET(loader.getHabit(id), response);
-});
-
-/* Get the info for the balance */
-router.add('GET', /^\/balance$/, function (request, response) {
-  simpleGET(loader.balance(), response);
-});
-
-/* Completes a habit: Adds the provided date array to the habit's log and
- * changes the balance appropriately */
-router.add('POST', /^\/complete-habit$/, function (request, response) {
-  const body = request.body;
-  const invalidMsg = validateRequest(body, ['id', 'rev', 'date', 'set']);
-  if (invalidMsg) {
-    respondBadReq(response, invalidMsg);
-  } else {
-    loader.updateDoc(body.id, body.rev, (origDoc) => {
-      let habit = Habit.fromDoc(origDoc);
-      if (body.set) habit.complete(body.date);
-      else habit.uncomplete(body.date);
-      return Object.assign(origDoc, habit.toDoc());
-    }).then(function (result) {
-      /* Get latest updates to the habit doc and balance */
-      return Promise.all([loader.getHabit(body.id), loader.balance()]);
-    }).then(function (docs) {
-      return response.end(JSON.stringify({
-        habit: docs[0],
-        balance: docs[1].balance
-      }));
-    }).catch(function (err) {
-      console.log(err);
-      response.statusCode = 400;
-      return response.end(JSON.stringify(err));
-    });
-  }
-});
-
-router.add('POST', /^\/charge-expense$/, function (request, response) {
-  const body = request.body;
-  const invalidMsg = validateRequest(body, ['id', 'rev', 'dateCharged']);
-  if (invalidMsg) {
-    respondBadReq(response, invalidMsg);
-  } else {
-    loader.updateDoc(body.id, body.rev, (origDoc) => {
-      let expense = Expense.fromDoc(origDoc);
-      if (body.dateCharged) expense.charge(body.dateCharged);
-      else expense.uncharge();
-      return Object.assign(origDoc, expense.toDoc());
-    }).then(function(result) {
-      /* Get latest updates */
-      return Promise.all([loader.getExpense(body.id), loader.balance()]);
-    }).then(function(result) {
-      return response.end(JSON.stringify({
-        expense: result[0],
-        balance: result[1].balance
-      }));
-    }).catch(function(err) {
-      console.log(err);
-      response.statusCode = 400;
-      return response.end(JSON.stringify(err));
-    });
-  }
-});
-
-/* TODO: FIX THIS -- A REV MUST BE USED HERE TO CONTROL CONCURRENCY */
-router.add('POST', /^\/change-balance$/, function (request, response) {
-  const body = request.body;
-  const invalidMsg = validateRequest(body, ['amount'], null, (b) => {
-    if (isNaN(b.amount)) return 'amount must be a valid number';
-  });
-  if (invalidMsg) {
-    respondBadReq(response, invalidMsg);
-  } else {
-    loader.getDoc('balance').then(function (doc) {
-      const changeAmt = Number(body.amount);
-      let balance = new Balance(doc.log);
-      balance.changeAmountBy(changeAmt);
-      return loader.updateDoc(Object.assign(doc, balance.toDoc()));
-    }).then(function (result) {
-      return loader.balance();
-    }).then(function (result) {
-      return response.end(JSON.stringify(result));
-    }).catch(function (err) {
-      response.statusCode = 400;
-      return response.end(JSON.stringify(err));
-    });
-  }
-});
-
-/* Update basic info about a habit -- but not the log -- never trust the client
- * Ignores everything in the request body except for the name and amount */
-router.add('POST', /^\/edit-habit$/, function (request, response) {
-  const body = request.body;
-  const invalidMsg = validateRequest(body, ['id', 'rev'], ['name', 'amount'],
-    (b) => {
-      if (typeof(b.name) === 'undefined' && typeof(b.amount) === 'undefined')
-        return 'either name, amount, or both must be provided';
-    }
-  );
-  if (invalidMsg) {
-    respondBadReq(response, invalidMsg);
-  } else {
-    loader.updateDoc(body.id, body.rev, (origDoc) => {
-      let habit = Habit.fromDoc(origDoc);
-      if (body.name) habit.name = body.name;
-      if (body.amount) habit.amount = body.amount;
-      return Object.assign(origDoc, habit.toDoc());
-    }).then(function (result) {
-      return loader.getHabit(body.id);
-    }).then(function (doc) {
-      response.end(JSON.stringify(doc));
-    }).catch(function (err) {
-      console.log(err);
-      response.statusCode = 400;
-      response.end(JSON.stringify(err));
-    });
-  }
-});
-
-router.add('POST', /^\/edit-expense/, function (request, response) {
-  const body = request.body;
-  const invalidMsg = validateRequest(body, ['id', 'rev'], ['name', 'amount'],
-    (b) => {
-      if (typeof(b.name) === 'undefined' && typeof(b.amount) === 'undefined')
-        return 'either name, amount, or both must be provided';
-    }
-  );
-  if (invalidMsg) {
-    respondBadReq(response, invalidMsg);
-  } else {
-    loader.updateDoc(body.id, body.rev, (origDoc) => {
-      let expense = Expense.fromDoc(origDoc)
-      if (body.name) expense.name = body.name;
-      if (body.amount) expense.amount = body.amount;
-      return Object.assign(origDoc, expense.toDoc());
-    }).then(function (result) {
-      /* This request can change the balance if the expense was completed and
-       * its amount changed */
-      return Promise.all([loader.getExpense(body.id), loader.balance()]);
-    }).then(function (docs) {
-      response.end(JSON.stringify({
-        expense: docs[0],
-        balance: docs[1]
-      }));
-    }).catch(function (err) {
-      console.log(err);
-      response.statusCode = 400;
-      response.end(JSON.stringify(err));
-    });
-  }
-});
-
-router.add('POST', /^\/archive-habit$/, function (request, response) {
-  const body = request.body;
-  const invalidMsg = validateRequest(body, ['id', 'rev', 'archived']);
-  if (invalidMsg) {
-    respondBadReq(response, invalidMsg);
-  } else {
-    loader.updateDoc(body.id, body.rev, (origDoc) => {
-      return Object.assign(origDoc, { archived: body.archived });
-    }).then(function (result) {
-      return loader.getHabit(body.id);
-    }).then(function (result) {
-      response.end(JSON.stringify(result));
-    }).catch(function (err) {
-      response.statusCode = 400;
-      response.end(JSON.stringify(err));
-    });
-  }
-});
-
-/* TODO: DRY out these create routes */
-/* Creates a new habit */
-router.add('PUT', /^\/habit$/, function (request, response) {
-  const body = request.body;
-  const invalidMsg = validateRequest(body, ['name', 'amount']);
-  if (invalidMsg) {
-    respondBadReq(response, invalidMsg);
-  } else {
-    const habit = new Habit(body.name, body.amount);
-    loader.createHabit(habit).then(function(result) {
-      return loader.getHabit(result.id);
-    }).then(function (result) {
-      response.statusCode = 200;
-      response.end(JSON.stringify(result));
-    }).catch(function (err) {
-      response.statusCode = 400;
-      response.end(JSON.stringify(err));
-    });
-  }
-});
-
-/* Creates a new expense */
-router.add('PUT', /^\/expense$/, function (request, response) {
-  const body = request.body;
-  const invalidMsg = validateRequest(body, ['name', 'amount']);
-  if (invalidMsg) {
-    respondBadReq(response, invalidMsg);
-  } else {
-    const expense = new Expense(body.name, body.amount);
-    loader.createExpense(expense).then(function(result) {
-      return loader.getExpense(result.id);
-    }).then(function (result) {
-      response.statusCode = 200;
-      response.end(JSON.stringify(result));
-    }).catch(function (err) {
-      response.statusCode = 400;
-      response.end(JSON.stringify(err));
-    });
-  }
-});
-
-router.routes = router.routes.concat(require('./routes/choreRoutes').routes);
 http.createServer(function (request, response) {
   let body = [];
   request.on('data', function (chunk) {
